forge fmt

+++ src/Project.sol
@@ -27,23 +27,17 @@ contract Project is IProject {
 
     function _refund(address payable to, uint256 amount) internal {
         if (amount == 0) return;
-        (bool ok, ) = to.call{value: amount}("");
+        (bool ok,) = to.call{value: amount}("");
         require(ok, "Refund failed");
     }
 
     modifier onlyProjectManager() {
-        require(
-            msg.sender == projectManager,
-            "Only the project manager can execute this function."
-        );
+        require(msg.sender == projectManager, "Only the project manager can execute this function.");
         _;
     }
 
     modifier onlyCreator() {
-        require(
-            msg.sender == creator,
-            "Only the project creator can execute this function."
-        );
+        require(msg.sender == creator, "Only the project creator can execute this function.");
         _;
     }
 
@@ -78,10 +72,7 @@ contract Project is IProject {
 
     // Función para actualizar el estado del proyecto
     function updateState(IProject.ProjectState _newState) external onlyProjectManager {
-        require(
-            uint(_newState) > uint(currentState),
-            "New state must be a higher phase."
-        );
+        require(uint256(_newState) > uint256(currentState), "New state must be a higher phase.");
         currentState = _newState;
         emit StateChanged(_newState);
     }
@@ -115,22 +106,15 @@ contract Project is IProject {
 
     // Función para comprar tokens con ETH
     function buyCarbonCredits(uint256 _amount) external payable override {
-
         // Verificar que el usuario haya enviado suficiente ETH
         uint256 totalCost = _amount * pricePerToken;
         require(msg.value >= totalCost, "Insufficient ETH sent");
 
         // Verificar que hay suficientes tokens liberados
-        require(
-            _amount <= getAvailableTokens(),
-            "Amount exceeds available tokens for this phase"
-        );
+        require(_amount <= getAvailableTokens(), "Amount exceeds available tokens for this phase");
 
         // Verificar que el contrato tiene suficientes tokens
-        require(
-            token.balanceOf(address(this)) >= _amount,
-            "Insufficient token balance"
-        );
+        require(token.balanceOf(address(this)) >= _amount, "Insufficient token balance");
 
         // Transferir tokens al usuario
         require(token.transfer(msg.sender, _amount), "Token transfer failed");
@@ -146,7 +130,7 @@ contract Project is IProject {
     // Función para que el creator retire el ETH acumulado
     function withdrawETH(uint256 _amount) public onlyCreator {
         require(address(this).balance >= _amount, "Insufficient ETH balance");
-        (bool ok, ) = payable(creator).call{value: _amount}("");
+        (bool ok,) = payable(creator).call{value: _amount}("");
         require(ok, "ETH transfer failed");
         emit ETHWithdrawn(creator, _amount);
     }
@@ -156,10 +140,7 @@ contract Project is IProject {
         uint256 totalCost = amount * pricePerToken;
         require(msg.value >= totalCost, "Insufficient ETH");
 
-        require(
-            getReleasedTokens() - purchasedTokens >= amount,
-            "Not enough tokens released"
-        );
+        require(getReleasedTokens() - purchasedTokens >= amount, "Not enough tokens released");
         require(token.balanceOf(address(this)) >= amount, "Insufficient balance");
 
         require(token.transfer(buyer, amount), "Transfer failed");
